+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Make it clear that `--profile` argument can be an HTTP(S) URL (and not only a path) (#288)
 - Fix README imprecisions + add back warc2zim availability in docker image (#314)
 - Enhance integration test to assert final content of the ZIM (#287)
+- Stop fetching and passing browsertrix crawler version as scraperSuffix to warc2zim (#354)
 
 ## [2.0.6] - 2024-08-02